+++ src/main/kotlin/org/ratelog/user/profile/ProfileController.kt
@@ -41,10 +41,8 @@ class ProfileController(
                     model.addAttribute("currentLang", it.lang.name)
                     model.addAttribute("isFollowed", it.isFollowed)
                     model.addAttribute("isLoggedUser", it.userId == user.id)
-                    model.addAttribute("movieRatings", it.movieRatings)
-                    model.addAttribute("tvRatings", it.tvRatings)
-                    model.addAttribute("movieHasMore", it.movieHasMore)
-                    model.addAttribute("tvHasMore", it.tvHasMore)
+                    model.addAttribute("feed", it.feed)
+                    model.addAttribute("hasMore", it.hasMore)
                     model.addAttribute("limit", limit)
                     "profile"
                 }